% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{AllScores}
\alias{AllScores}
\title{Compute All PWM Scores}
\usage{
AllScores(PWMs, dataSeq, workers = 4)
}
\arguments{
\item{PWMs}{A list of Position Weight Matrices (PWMs). Each PWM in the list
should be a numeric matrix with 4 rows (corresponding to A, C, G, T).}

\item{dataSeq}{A matrix containing chromosomes names and the DNA sequences to be
scored. Each sequence must consist of the characters "A", "C", "G", and "T".
The names of the vector will be used as row names in the output.}

\item{workers}{integer, the number of cores to assign for parallel computations}
}
\value{
A numeric matrix where:
\item{Rows}{Correspond to the DNA sequences (named from \code{dataSeq}).}
\item{Columns}{Correspond to each PWM in \code{PWMs}.}
Each entry is the maximum score for the given sequence and PWM.
}
\description{
This function calculates the maximum scores for a list of Position Weight
Matrices (PWMs) across a set of DNA sequences using parallel computation.
For each PWM, it computes the maximum score for every sequence and returns
a matrix of scores.
}
