% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata_documentation.R
\name{C.elegans}
\alias{C.elegans}
\title{Example Data for Caenorhabditis elegans}
\source{
Genome data for \emph{Caenorhabditis elegans} obtained from NCBI.
}
\description{
This dataset includes genomic data for Caenorhabditis elegans, consisting of a nucleotide sequence file and a genomic annotation file.
}
\section{Files}{

\describe{
\item{\code{GCF_000002985.6_WBcel235_genomic.fna}}{FASTA file containing nucleotide sequences for the Caenorhabditis elegans genome.}
\item{\code{c_elegans_genomic.gff}}{GFF file containing annotations for the genome, including gene positions and features.}
}
}

