% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_motifs.R
\name{consensus_fisher_test}
\alias{consensus_fisher_test}
\title{Consensus Fisher Test}
\usage{
consensus_fisher_test(consensus, num_of_regressions, alpha = 0.05)
}
\arguments{
\item{consensus}{A list where each element is a named vector
representing motif frequencies for different conditions.}

\item{num_of_regressions}{The total number of regressions
or occurrences to normalize motif frequencies.}

\item{alpha}{Significance level for the test, default is 0.05.}
}
\value{
A Fisher's test result object
containing p-values and test statistics.
}
\description{
This function performs a Fisher's exact test
on motif frequencies across multiple conditions.
The input is a consensus data.frame,
which contains the motif frequencies for each condition.
}
