% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_motifs.R
\name{consensus_motifs}
\alias{consensus_motifs}
\title{Consensus Motifs}
\usage{
consensus_motifs(df_list, by_regressor = TRUE)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{by_regressor}{Logical. If \code{TRUE}, the consensus motifs are
calculated by condition across all data frames. If \code{FALSE}, the
consensus motifs are calculated for each data frame individually.}
}
\value{
A named list where each element contains the consensus motifs as a
named vector of percentages.
}
\description{
This function calculates the consensus motifs across a list of data frames.
}
