% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_motifs.R
\name{consensus_motifs_heatmap}
\alias{consensus_motifs_heatmap}
\title{Heatmap of Consensus Motif Inclusivity}
\usage{
consensus_motifs_heatmap(consensus_output)
}
\arguments{
\item{consensus_output}{A named list generated by the \code{consensus_motifs}
function. The names of the list represent either conditions or regression
models, and the elements contain named vectors of motif inclusivity
percentages.}
}
\value{
A ComplexHeatmap object representing the heatmap.
}
\description{
This function creates a heatmap to visualize the inclusivity of motifs
across conditions or regression models based on the output of the
\code{consensus_motifs} function.
}
