% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata_documentation.R
\name{D.melanogaster}
\alias{D.melanogaster}
\title{Example Data for Drosophila melanogaster}
\source{
Genome data for \emph{Drosophila melanogaster} obtained from NCBI.
}
\description{
This dataset includes genomic data for Drosophila melanogaster, consisting of a nucleotide sequence file and a genomic annotation file.
}
\section{Files}{

\describe{
\item{\code{GCF_000001215.4_Release_6_plus_ISO1.fna}}{FASTA file containing nucleotide sequences for the Drosophila melanogaster genome.}
\item{\code{d_melanogaster_genomic.gff}}{GFF file containing annotations for the genome, including gene positions and features.}
}
}

