% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata_documentation.R
\name{E.coli}
\alias{E.coli}
\title{Example Data for Escherichia coli}
\source{
Genome data for \emph{Escherichia coli} obtained from the coordinator of the project and NCBI.
}
\description{
This dataset includes genomic data for Escherichia coli, consisting of a nucleotide sequence file, a genomic annotation file,
and additional data files related to transcription units, protein information, and motifs.
}
\section{Files}{

\describe{
\item{\code{GCF_000005845.2_ASM584v2_genomic.fna}}{FASTA file containing nucleotide sequences for the Escherichia coli genome.}
\item{\code{GCF_000005845.2_ASM584v2_genomic.gff}}{GFF file containing annotations for the genome, including gene positions and features.}
\item{\code{compendium.xlsx}}{Excel file containing Transcripts Per Million (TPM) for different genes under certain conditions.}
\item{\code{511145_protein_info_v12_0.txt}}{Txt file containing information about E. coli protein, like protein_id, protein name, protein size, and annotation.}
\item{\code{SwissRegulon_e_coli.meme}}{Meme file containing Position Frequency Matrices (PFMs) for various transcription factor motifs.}
\item{\code{TUSet.tsv}}{Tsv file containing tabular dataset of transcription units from Escherichia coli K-12, sourced from RegulonDB 13.0. It contains detailed information about transcription units, their associated operons, genes, promoters, terminators, evidence, and confidence levels.}
}
}

