% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{extract_sequences}
\alias{extract_sequences}
\title{Extract DNA Sequences from a FASTA Object}
\usage{
extract_sequences(FASTA, row)
}
\arguments{
\item{FASTA}{A named list or vector of DNA sequences, where each element is
a sequence.}

\item{row}{A vector containing the chromosome name, start position, and
end position of the sequence to extract. The row should be in the
format: c(chr, start, end).}
}
\value{
A character vector with three elements:
\item{chromosome}{The name of the chromosome from which the sequence is
extracted.}
\item{gene_name}{The gene name, extracted from the row names.}
\item{sequence}{The extracted DNA sequence.}
}
\description{
This function extracts a DNA sequence from a FASTA object based on the
provided chromosome, start, and end positions. It returns the sequence along
with its associated chromosome and gene name.
}
