% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{get_scores}
\alias{get_scores}
\title{Compute PWM Scores}
\usage{
get_scores(PWM, dataSeq)
}
\arguments{
\item{PWM}{A numeric matrix representing the Position Weight Matrix (PWM).
The matrix should have 4 rows (for A, C, G, T).}

\item{dataSeq}{A character vector containing DNA sequences to be scored.
Each sequence must be composed of the characters "A", "C", "G", and "T".}
}
\value{
A numeric matrix with two columns and a number of rows equal to the
length of \code{dataSeq}:
\item{Column 1}{The position in the sequence with the highest score.}
\item{Column 2}{The maximum score for the sequence.}
}
\description{
This function calculates the maximum position-specific scoring matrix (PSSM)
scores for a list of DNA sequences using a given Position Weight Matrix(PWM).
It identifies the position with the highest score for each sequence and
returns both the position and the corresponding score.
}
