% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFM_loader.R
\name{local_PFM_loader}
\alias{local_PFM_loader}
\title{local_PFM_loader}
\usage{
local_PFM_loader(path)
}
\arguments{
\item{path}{A character string representing the path to the file containing
PFMs.The file should follow the MEME motif format, with motifs identified by
the keyword "MOTIF".}
}
\value{
A named list of data frames. Each data frame corresponds to the PFM
of a motif and contains the nucleotide frequencies at each position
("A", "C", "G", "T").The names of the list elements are the motif
names extracted from the file.
}
\description{
This function reads a file containing position frequency matrices (PFMs) and
extracts them into a list of data frames. Each PFM represents the nucleotide
frequencies ("A", "C", "G", "T") at each position for a given motif.
}
