% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_first_genes.R
\name{match_genes}
\alias{match_genes}
\title{Match Genes to Protein Information}
\usage{
match_genes(tuGenes_vector, protein_info_data)
}
\arguments{
\item{tuGenes_vector}{A character vector containing the gene names to match.}

\item{protein_info_data}{A data frame where the first column contains
protein IDs and the second column contains gene names.}
}
\value{
A data frame containing two columns: the matched protein IDs (from
the first column) and the preferred names (from the second column) for the
input gene names.
}
\description{
This function matches a vector of gene names to a protein information
dataset and returns the corresponding protein IDs and preferred names.
}
