% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2dataframe.R
\name{models2dataframe}
\alias{models2dataframe}
\title{Convert Regression Models to Dataframes}
\usage{
models2dataframe(regression_list, k = 5, alpha = NULL)
}
\arguments{
\item{regression_list}{A list of regression models. The names of the list
elements should correspond to the model types ("Randomized_lasso", "PLS",
"SVM", "LM", "KNN","RF).}

\item{k}{An integer specifying the number of top features to include for
unpacking. Default is 5.}

\item{alpha}{A numeric value specifying the significance threshold for
linear models (required only for LM). Default is NULL.}
}
\value{
A list of data frames, each corresponding to a model type and
containing unpacked results for each condition.
}
\description{
This function processes a list of regression models (Randomized Lasso, PLS,
SVM, and LM) and converts them into a data frame for further analysis.
The models are unpacked based on their type and conditions.
}
