% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{MotifRegressR}
\alias{MotifRegressR-package}
\alias{MotifRegressR}
\title{MotifRegressR: Perform motif regression on gene expression conditions}
\description{
This R package facilitates motif analysis
for gene expression studies. It integrates upstream DNA sequence information,
gene expression profiles, and Position Frequency Matrices (PFMs)
to identify regulatory motifs influencing gene expression.
Users can import PFMs from JASPAR or custom sources,
annotate genomic sequences with transcription start sites (TSS),
and retrieve metadata of gene expression conditions.
The package scores motifs across genes, aggregates the results,
and performs regression analysis to determine the most impactful motifs
under different experimental conditions. Visualization functions finally
help the user in the sense-making step of the analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeonardoNossa/Motif-regressor}
  \item Report bugs at \url{https://github.com/LeonardoNossa/Motif-regressor/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Nossa \email{leonardo.nossa@studenti.unimi.it}

Authors:
\itemize{
  \item Elena Sasso
  \item Gabriele Oliveto
  \item Marco Cominelli
  \item Teodor Devda
  \item Riccardo de Sury
  \item Laura Mottarlini
  \item Federico Camozzi
}

}
\keyword{package}
