% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFM2PWM.R
\name{PFM2PWM}
\alias{PFM2PWM}
\title{Convert Position Frequency Matrices(PFMs) to Position Weight Matrices(PWMs)}
\usage{
PFM2PWM(PFMs, background = c(0.25, 0.25, 0.25, 0.25))
}
\arguments{
\item{PFMs}{A named list of data frames, where each data frame represents a
Position Frequency Matrix (PFM) for a motif. Each PFM should contain
nucleotide frequencies for "A", "C", "G", and "T" at each position.}

\item{background}{A numeric vector of length 4 representing the background
nucleotide probabilities for "A", "C", "G", and "T".
The default is \code{c(0.25, 0.25, 0.25, 0.25)}. Values should sum to 1.}
}
\value{
A named list of data frames, where each data frame corresponds to a
Position Weight Matrix (PWM) for a motif. Each PWM contains log2-transformed
weights relative to the background probabilities.
}
\description{
This function takes a list of Position Frequency Matrices (PFMs) and converts
them into Position Weight Matrices (PWMs) using the given background
nucleotide frequencies.
}
