% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustermap.R
\name{pivot_df}
\alias{pivot_df}
\title{Pivot data.frame}
\usage{
pivot_df(df)
}
\arguments{
\item{df}{the data.frame coming from \code{models2dataframe}, containing the top
k motif in each condition and their importance score}
}
\value{
A data.frame, where rows are motifs, and columns are the conditions.
If a motif is not among the top k motifs in that condition, the function will
assign 0.
}
\description{
This function takes as input a data.frame coming
from \code{models2dataframe} call and pivots it, assigning the importance score
for each motif, in each condition.
}
