% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustermap.R
\name{plot_clustermap}
\alias{plot_clustermap}
\title{Plot clustermap}
\usage{
plot_clustermap(
  df_sublist,
  distance = c("euclidean", "euclidean"),
  method = c("complete", "complete")
)
}
\arguments{
\item{df_sublist}{A sublist coming from \code{models2dataframe}}

\item{distance}{a vector of length 2, containing the distance metrics to be
used. Each element can be a pre-defined character which is in
("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski",
"pearson", "spearman", "kendall"). It can also be a function.
If the function has one argument, the input argument should be a matrix
and the returned value should be a dist object.
If the function has two arguments, the input arguments are two vectors
and the function calculates distance between these two vectors.}

\item{method}{a vector of length 2, containing the clustering method to
perform. the agglomeration method to be used. This should be one of "ward.D",
"ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC).}
}
\value{
a clustermap, showing the grouping of motifs and conditions, based on
the importance values.
}
\description{
This function takes as input A sublist coming from \code{models2dataframe},
distances metrics and clustering approaches. The data.frame inside the
sublist is first pivoted using \code{pivot_df} function, then a clustermap is
built
}
