% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_LM}
\alias{plot_LM}
\title{Plot Linear Model (LM) Diagnostics and Coefficients}
\usage{
plot_LM(LM_model)
}
\arguments{
\item{LM_model}{A trained linear model object.}
}
\value{
A list containing data frames for
coefficients and recorded plots for diagnostic analysis.
}
\description{
Generates diagnostic plots and
coefficients data for a Linear Model (LM).
}
