% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_O2PLS}
\alias{plot_O2PLS}
\title{Plot O2PLS Model Metrics and Feature Importance}
\usage{
plot_O2PLS(O2PLS_model)
}
\arguments{
\item{O2PLS_model}{A trained O2PLS model object.}
}
\value{
A list containing motifs weights and ggplot objects
for such weights and correlation between joint X scores and joint Y scores.
}
\description{
Generates plots and data summaries for
O2PLS model metrics (motifs weights and correlation between joint X scores
and joint Y scores).
}
