% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_PLS}
\alias{plot_PLS}
\title{Plot Partial Least Squares (PLS) Model Coefficients and Validation}
\usage{
plot_PLS(PLS_model)
}
\arguments{
\item{PLS_model}{A trained PLS model object from the pls package.}
}
\value{
A list containing ggplot/plotly plots
and data frames for PLS coefficients and validation.
}
\description{
Generates component coefficient plots, heatmaps,
and validation plots for a PLS model.
}
