% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_RANDLASSO}
\alias{plot_RANDLASSO}
\title{Plot Random LASSO Model Stability and Selection Probability}
\usage{
plot_RANDLASSO(RANDLASSO_model)
}
\arguments{
\item{RANDLASSO_model}{A trained Random LASSO model
object from the monaLisa package.}
}
\value{
A list containing stability paths,
selection probability plots, and selection probability data.
}
\description{
Generates stability paths and selection
probability plots for a Random LASSO model.
}
