% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_SVM}
\alias{plot_SVM}
\title{Plot SVM Model Feature Importance and Metrics}
\usage{
plot_SVM(SVM_model_sublist, Compendium, Scores)
}
\arguments{
\item{SVM_model_sublist}{A list containing SVM model objects.}

\item{Compendium}{A data frame with actual values.}

\item{Scores}{A data frame with predicted scores.}
}
\value{
A list containing a ggplot object for feature importance,
a data frame of feature importance, and a metrics data frame.
}
\description{
Generates feature importance bar plot and metrics
(MSE, R2, RMSE, MAE) for an SVM model.
}
