% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_lasso}
\alias{Regression_lasso}
\title{Randomized Lasso Regression}
\usage{
Regression_lasso(Scores, TPMs, Condition, params)
}
\arguments{
\item{Scores}{A matrix or data frame of scores to be used as predictors in
the Randomized Lasso regression. This is internally converted to a matrix.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}
}
\value{
An object returned by the \code{randLassoStabSel} function,
representing the results of the stability selection process.
}
\description{
This function performs Randomized Lasso regression for stability selection
using the \code{randLassoStabSel} function from the \code{monaLisa} package.
}
