% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_O2PLS}
\alias{Regression_O2PLS}
\title{O2PLS Regression}
\usage{
Regression_O2PLS(Scores, TPMs, Condition, params)
}
\arguments{
\item{Scores}{A matrix or data frame of scores to be used as predictors in
the PLS regression. This is internally converted to a matrix.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}

\item{params}{A named vector containing optional parameters}
}
\value{
A list representing the O2PLS regression
}
\description{
This function performs O2PLS regression using the
\code{o2m} function from the \code{OmicsPLS} package.
}
