% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_pls}
\alias{Regression_pls}
\title{Partial Least Squares (PLS) Regression}
\usage{
Regression_pls(Scores, TPMs, Condition, params)
}
\arguments{
\item{Scores}{A matrix or data frame of scores to be used as predictors in
the PLS regression. This is internally converted to a matrix.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}
}
\value{
An object of class \code{mvr} returned by the \code{plsr} function,
representing the PLS regression model with the optimal number of components.
}
\description{
This function performs Partial Least Squares (PLS) regression using the
\code{plsr} function from the \code{pls} package. It determines the optimal
number of components using cross-validation and builds the final model
accordingly.
}
