% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_RF}
\alias{Regression_RF}
\title{Random Forest Regression}
\usage{
Regression_RF(Scores, TPMs, Condition, params)
}
\arguments{
\item{Scores}{A matrix or data frame of scores to be used as predictors in
the Random Forest regression. This is converted to a matrix internally.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}
}
\value{
An object of class \code{randomForest} returned by the
\code{randomForest} function, representing the fitted Random Forest
regression model.
}
\description{
This function performs Random Forest regression using the
\code{randomForest} function from the \code{randomForest} package.
}
