% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_simple}
\alias{Regression_simple}
\title{Simple Linear Regression}
\usage{
Regression_simple(Scores, TPMs, Condition, params)
}
\arguments{
\item{Scores}{A data frame or matrix of scores to be used as predictors in
the regression model.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}
}
\value{
An object of class \code{lm} returned by the \code{lm} function,
representing the fitted linear model.
}
\description{
This function performs a simple linear regression using the \code{lm}
function.
}
