% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{Regression_svm}
\alias{Regression_svm}
\title{Support Vector Machine (SVM) Regression}
\usage{
Regression_svm(Scores, TPMs, Condition)
}
\arguments{
\item{Scores}{A matrix or data frame of scores to be used as predictors in
the SVM regression. This is internally converted to a matrix.}

\item{TPMs}{A data frame or matrix of TPM (Transcripts Per Million) values.
Columns correspond to different conditions or samples.}

\item{Condition}{A character string specifying the column name in \code{TPMs}
corresponding to the condition to be analyzed.}
}
\value{
An object of class \code{svm} returned by the \code{svm} function,
representing the fitted SVM regression model.
}
\description{
This function performs SVM regression using the \code{svm} function from
the \code{e1071} package with a linear kernel.
}
