% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{regressor}
\alias{regressor}
\title{Regressor}
\usage{
regressor(
  scores,
  compendium,
  conditions,
  regression_type,
  regression_params_list
)
}
\arguments{
\item{scores}{A data frame or matrix of scores (e.g., feature scores or
experimental data).}

\item{compendium}{A data frame or matrix of the compendium data (e.g.,
known conditions or targets for regression).}

\item{conditions}{A vector of conditions or responses used for training the
regression models.}

\item{regression_type}{A character vector specifying the types of regression
models to be trained. Available types depend on the \code{train_models} function.}
}
\value{
A list of regression models, each corresponding to a specified
regression type.
}
\description{
This function trains multiple regression models on the provided scores and
compendium data. It filters the rows to retain only the common ones between
the scores and the compendium before training the models.
}
