% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_frequencies.R
\name{retrieve_frequencies}
\alias{retrieve_frequencies}
\title{Obtain the nucleotide frequencies from a FASTA file}
\usage{
retrieve_frequencies(fasta)
}
\arguments{
\item{fasta}{The path to the input FASTA file}
}
\value{
A named vector with frequencies for each nucleotide
(in this order A,C,G,T)
}
\description{
This function takes as input FASTA file and computes the frequencies of each
nucleotide.
}
