% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metadata.R
\name{retrieve_metadata}
\alias{retrieve_metadata}
\title{Retrieve Compendium Metadata}
\usage{
retrieve_metadata(compendium)
}
\arguments{
\item{compendium}{A matrix where each row is a gene and each column
is a sample. The names of the columns should be IDs retrievable from ENA
browser (e.g. SRR, ERR, DRR etc...). The user is notified when an ID is not
retrievable. An error is displayed when no ID is retrievable}
}
\value{
A data.frame, where each row is a sample and columns are metadata
}
\description{
This function takes as input a gene expression compendium
and uses the sample IDs to retrieve metadata from ENA browser
}
