% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_motifs.R
\name{retrieve_regressor_type}
\alias{retrieve_regressor_type}
\title{Retrieve the Type of a Regression Model}
\usage{
retrieve_regressor_type(model)
}
\arguments{
\item{model}{An object representing a regression model. The model's class is
used to determine its type.}
}
\value{
A character string indicating the type of regression model:
\itemize{
\item \code{"svm"} for Support Vector Machine models.
\item \code{"pls"} for Partial Least Squares regression models.
\item \code{"lasso"} for Randomized Lasso regression models (stored as
\code{SummarizedExperiment} objects).
\item \code{"simple"} for simple linear regression models.
}
}
\description{
This function determines the type of a given regression model based on its
class.
}
