% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata_documentation.R
\name{S.cerevisiae}
\alias{S.cerevisiae}
\title{Example Data for Saccharomyces cerevisiae}
\source{
Genome data for \emph{Saccharomyces cerevisiae} obtained from the coordinator of the project and NCBI.
}
\description{
This dataset includes genomic data for Saccharomyces cerevisiae, consisting of a nucleotide sequence file, a genomic annotation file, and gene expression data
}
\section{Files}{

\describe{
\item{\code{GCA_000146045.2_R64_genomic.fna}}{FASTA file containing nucleotide sequences for the Saccharomyces cerevisiae genome.}
\item{\code{GCA_000146045.2_R64_genomic.gff}}{GFF file containing annotations for the genome, including gene positions and features.}
}
}

