% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_first_genes.R
\name{table_first_genes}
\alias{table_first_genes}
\title{Create a Table of Filtered Genes and Protein Information}
\usage{
table_first_genes(tsv, pro)
}
\arguments{
\item{tsv}{A string specifying the path to the TSV file containing gene
information.}

\item{pro}{A string specifying the path to the file containing protein
information.}
}
\value{
A data frame with two columns: \code{Gene_ID} (the matched protein ID)
and \code{Gene_Name} (the corresponding gene name).
}
\description{
This function reads a TSV file of genes and a second file with protein
information, filters the genes, matches them to the protein data, and
create the resulting table.
}
