% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{train_models}
\alias{train_models}
\title{Train Regression Models}
\usage{
train_models(S, EXPN, conditions, regression, regression_params)
}
\arguments{
\item{S}{A numeric matrix or data frame of scores}

\item{EXPN}{A numeric matrix or data frame of expression values}

\item{conditions}{A vector specifying the conditions to be analyzed. Can be
column indices or names in \code{EXPN}. Use \code{"all"} to include all columns.}

\item{regression}{A character string indicating the regression method to use.
Options are:
\itemize{
\item \code{"Randomized_lasso"} for Lasso regression with stability selection.
\item \code{"PLS"} for Partial Least Squares regression.
\item \code{"SVM"} for Support Vector Machine regression.
\item \code{"LM"} for simple linear regression.
\item \code{"KNN"} for KNN regression.
}}
}
\value{
A named list where each element corresponds to a regression model for
a specific condition.
}
\description{
This function trains regression models for specified experimental conditions
using the selected regression method.
}
