% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2dataframe.R
\name{unpack_lasso}
\alias{unpack_lasso}
\title{Unpack Lasso}
\usage{
unpack_lasso(lasso_sublist, k)
}
\arguments{
\item{lasso_sublist}{A list containing a single lasso regression model. The
model must include metadata with stability selection parameters.}

\item{k}{An integer specifying the number of top motifs to extract.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{motifs}: The names of the top motifs.
\item \code{conditions}: The name of the condition corresponding to the motifs.
\item \code{coef}: The stability selection probabilities (in percentages) of the
motifs.
}
If no motifs are selected, the function returns \code{NULL}.
}
\description{
This function extracts the top \code{k} motifs from a Randomized lasso
regression model based on their stability selection probabilities.
}
