% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2dataframe.R
\name{unpack_pls}
\alias{unpack_pls}
\title{Unpack PLS}
\usage{
unpack_pls(pls_sublist, k)
}
\arguments{
\item{pls_sublist}{A list containing a single PLS regression model. The model
must include coefficients for motifs.}

\item{k}{An integer specifying the number of top motifs to extract.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{motifs}: The names of the top motifs.
\item \code{conditions}: The name of the condition corresponding to the motifs.
\item \code{coef}: The regression coefficients of the motifs.
}
}
\description{
This function extracts the top \code{k} motifs from a PLS regression model based
on their regression coefficients.
}
