% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2dataframe.R
\name{unpack_RF}
\alias{unpack_RF}
\title{Unpack RF}
\usage{
unpack_RF(RF_sublist, k)
}
\arguments{
\item{RF_sublist}{A list containing a single RF model (object of
class \code{randomForest}).}

\item{k}{An integer specifying the number of top features to extract.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{motifs}: The names of the top motifs.
\item \code{conditions}: The name of the condition corresponding to the motifs.
\item \code{coef}: The feature importance values for the motifs.
}
}
\description{
This function extracts the top \code{k} most important motifs from a
trained RF model based on their feature importance.
}
