% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2dataframe.R
\name{unpack_simple}
\alias{unpack_simple}
\title{Unpack simple}
\usage{
unpack_simple(simple_sublist, k, alpha)
}
\arguments{
\item{simple_sublist}{A list containing a single linear regression model
(object of class \code{lm}).}

\item{k}{An integer specifying the number of top motifs to extract.}

\item{alpha}{A numeric value specifying the significance threshold for
p-values (e.g., 0.05).}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{motifs}: The names of the top motifs (excluding the intercept term).
\item \code{conditions}: The name of the condition corresponding to the motifs.
\item \code{coef}: The estimated coefficients of the motifs.
}
If no coefficients pass the p-value threshold \code{alpha}, the function returns
\code{NULL}.
}
\description{
This function extracts the top \code{k} motifs from a simple linear regression
model based on their p-values and coefficients.
}
