% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Extract DNA Sequences Based on GFF and FASTA Files}
\usage{
get_sequences(gff, fasta, is_circular, upstream = 300)
}
\arguments{
\item{gff}{A string specifying the path to the GFF file containing genomic
features and regions.}

\item{fasta}{A string specifying the path to the FASTA file containing the
DNA sequences.}

\item{is_circular}{A logical value indicating whether the chromosomes are
circular.}

\item{upstream}{An integer specifying the number of base pairs to include
upstream of each gene. Default is 300.}
}
\value{
A data frame with two columns:
\item{Chr}{The chromosome name for each extracted sequence.}
\item{Sequence}{The DNA sequence for each gene, extracted from the specified
region in the FASTA file.}
}
\description{
This function extracts DNA sequences based on gene regions defined in a GFF
file and retrieves the corresponding sequences from a FASTA file.
The function returns a data frame with the extracted sequences for each gene.
}
